// src/nodecg/generated/spotifyPlayback.d.ts
/* prettier-ignore */
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface SpotifyPlayback {
	isPlaying: boolean;
	track?: {
		id: string;
		name: string;
		uri?: string;
		artists?: string[];
		albumName?: string;
		albumImageUrl?: string;
		durationMs?: number;
	} | null;
	progressMs?: number;
	deviceName?: string;
	updatedAt?: string;
}