// src/nodecg/generated/spotifyStatus.d.ts
/* prettier-ignore */
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface SpotifyStatus {
	authorized: boolean;
	user?: {
		id: string;
		displayName: string;
	} | null;
	lastSynced?: string;
	lastError?: string;
}